# backend/My Project.openapi.json
{"openapi":"3.0.1","info":{"title":"My Project","description":"","version":"1.0.0"},"tags":[{"name":"pet"}],"paths":{"/pet/{petId}":{"get":{"summary":"Find pet by ID","deprecated":false,"description":"","tags":["pet"],"parameters":[{"name":"petId","in":"path","description":"pet ID","required":true,"example":"1","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer","minimum":0,"maximum":0,"description":"status code"},"data":{"$ref":"#/components/schemas/Pet","description":"pet details"}},"required":["code","data"]},"examples":{"1":{"summary":"Success example","value":{"code":0,"data":{"name":"Hello Kitty","photoUrls":["http://dummyimage.com/400x400"],"id":3,"category":{"id":71,"name":"Cat"},"tags":[{"id":22,"name":"Cat"}],"status":"sold"}}}}}}},"400":{"description":"Invalid input","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer"},"message":{"type":"string"}},"required":["code","message"]}}}},"404":{"description":"Record not found","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer"},"message":{"type":"string"}},"required":["code","message"]}}}}},"security":[]},"delete":{"summary":"Deletes a pet","deprecated":false,"description":"","tags":["pet"],"parameters":[{"name":"petId","in":"path","description":"Pet id to delete","required":true,"schema":{"type":"string"}},{"name":"api_key","in":"header","description":"","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer","minimum":0,"maximum":0}},"required":["code"]},"examples":{"1":{"summary":"Success example","value":{"code":0}}}}}}},"security":[]}},"/pet":{"post":{"summary":"Add a new pet to the store","deprecated":false,"description":"","tags":["pet"],"parameters":[],"requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"type":"object","properties":{"name":{"description":"Pet Name","example":"Hello Kitty","type":"string"},"status":{"description":"Pet Sales Status","example":"sold","type":"string"}},"required":["name","status"]}}}},"responses":{"201":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer","minimum":0,"maximum":0},"data":{"$ref":"#/components/schemas/Pet","description":"pet details"}},"required":["code","data"]},"examples":{"1":{"summary":"Success example","value":{"code":0,"data":{"name":"Hello Kitty","photoUrls":["http://dummyimage.com/400x400"],"id":3,"category":{"id":71,"name":"Cat"},"tags":[{"id":22,"name":"Cat"}],"status":"sold"}}}}}}}},"security":[]},"put":{"summary":"Update an existing pet","deprecated":false,"description":"","tags":["pet"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer"},"data":{"$ref":"#/components/schemas/Pet","description":"pet details"}},"required":["code","data"]},"examples":{"1":{"summary":"Success example","value":{"code":0,"data":{"name":"Hello Kitty","photoUrls":["http://dummyimage.com/400x400"],"id":3,"category":{"id":71,"name":"Cat"},"tags":[{"id":22,"name":"Cat"}],"status":"sold"}}}}}}},"404":{"description":"Record not found","content":{"application/json":{"schema":{"type":"object","properties":{}}}}},"405":{"description":"Validation error","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/api/product":{"get":{"summary":"Get Product List","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/api/product/1":{"get":{"summary":"Get Particular Product","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/pet/findByStatus":{"get":{"summary":"Finds Pets by status","deprecated":false,"description":"","tags":["pet"],"parameters":[{"name":"status","in":"query","description":"Status values that need to be considered for filter","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Pet","description":"pet details"}},"examples":{"1":{"summary":"Success example","value":{"code":0,"data":[{"name":"Hello Kitty","photoUrls":["http://dummyimage.com/400x400"],"id":3,"category":{"id":71,"name":"Cat"},"tags":[{"id":22,"name":"Cat"}],"status":"sold"},{"name":"White Dog","photoUrls":["http://dummyimage.com/400x400"],"id":3,"category":{"id":71,"name":"Dog"},"tags":[{"id":22,"name":"Dog"}],"status":"sold"}]}}}}}},"400":{"description":"Invalid status value","content":{"application/json":{"schema":{"type":"object","properties":{"code":{"type":"integer"}},"required":["code"]}}}}},"security":[]}},"/api/addproduct":{"post":{"summary":"Add Product","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{},"required":["01J4J58G099SFNXPZ8H1KV1VMG"]},"example":{"name":"XYZ","description":"HELLO WORLD","price":10.2,"stock":0}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/api/updateproduct":{"post":{"summary":"Update Product","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"Id":17,"Name":"ey","Description":"","Price":10,"Stock":0}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/api/deleteproduct/13":{"get":{"summary":"Delete Product","deprecated":false,"description":"","tags":[],"parameters":[],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}}},"components":{"schemas":{"Tag":{"type":"object","properties":{"id":{"type":"integer","format":"int64","minimum":1,"description":"Tag ID"},"name":{"type":"string","description":"Tag Name"}}},"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64","minimum":1,"description":"Category ID"},"name":{"type":"string","description":"Category Name"}}},"Pet":{"required":["name","photoUrls","id","category","tags","status"],"type":"object","properties":{"id":{"type":"integer","format":"int64","minimum":1,"description":"Pet ID"},"category":{"$ref":"#/components/schemas/Category","description":"group"},"name":{"type":"string","description":"name","examples":["doggie"]},"photoUrls":{"type":"array","items":{"type":"string"},"description":"image URL"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"},"description":"tag"},"status":{"type":"string","description":"Pet Sales Status","enum":["available","pending","sold"]}}}},"securitySchemes":{}},"servers":[]}